Added HeroButton and Button component
Added HeroButton and Button component as they are designed differently.
Apart from that I set up a payment via stripe on all my events. When clicked on the button it will redirect the user to the payment page. User can enter a 4242 4242 4242 4242 card number and dummy expiry date/cvv/name and payment will be successful.